Update and rename CtgService.cs to Service.cs
